Moved Group exception classes to group.exceptions package
