+++ src/main/java/seedu/address/model/group/exceptions/GroupNotFoundException.java
@@ -1,4 +1,4 @@
-package seedu.address.model.person.exceptions;
+package seedu.address.model.group.exceptions;
 
 /**
  * Signals that the operation is unable to find the required Group
